Weather and Wind | Physics
In this small project that i worked on, i created a rain particle system and used 3 different scenes, Sunny , Cloudy and Rainy.  The wind and rain varies in each scene. I also built in a small feature where once you click your mouse - A shelter will be created, this will normalize the velocity (if the ball is in the shelter).