Record and play
Record and play is a small replay system which records a set of movements which can be saved and replayed.
To create a movement file, just type recordReplay() in the draw method. Once you are done, press 'e' to exit. To load your movements, type loadReplay() in the draw function. This will replicate everything the player did.